Separate tests for `ember addon` from tests for `ember new`
